Fix a bug in Azure config parsing.
Previously, if the id-file field was present in the config, but set to an empty string, we would handle it incorrectly, and assume that the id file was the folder "~/.ssh". Fix this to make our logic resilient to both the field being set to an empty string, and the field not being present.