Add version switch for yq 4
The dev for yq has completely reworked the syntax for version 4.
As we're using homebrew to install it in `setup.sh`, it always installs
the newest version. This will break quickstyle, as it calls yq with the
(now) incorrect syntax.
The version switch ensures that yq always gets called with the right
syntax. Quickstyle will also print console info hinting the user to
upgrade yq to > 4.0.0.